添加播放列表显示功能

+++ app/src/main/java/com/language/repeater/pcm/WaveformPoint.kt
@@ -1,3 +1,3 @@
 package com.language.repeater.pcm
 
-data class WaveformPoint(val time: Float, val min: Int, val max: Int)
+data class WaveformPoint(val time: Float, val min: Short, val max: Short)
